+++ footlog/src/api/home/search/deleteRecentSearch.ts
@@ -1,6 +1,6 @@
 import api from 'api/api';
 import { Response } from 'types/common/Response';
-import { SearchLogDtoDataTypes } from 'types/search/SearchTypes';
+import { SearchLogDtoDataTypes } from 'types/home/search/SearchTypes';
 
 export async function deleteRecentSearch(props: SearchLogDtoDataTypes): Promise<Response<any>> {
   const { keyword } = props;
