+++ footlog/src/api/home/search/getCityRegions.ts
@@ -1,11 +1,6 @@
 import api from 'api/api';
 import { Response } from 'types/common/Response';
-
-export interface CityRegionsDtoDataTypes {
-  sigunguCode: number;
-  areaCode: number;
-  sigunguName: string;
-}
+import { CityRegionsDtoDataTypes } from 'types/home/search/SearchTypes';
 
 export async function getCityRegions(): Promise<Response<CityRegionsDtoDataTypes[]>> {
   const { data } = await api.get(`/search/sigungu_code`);
