+++ footlog/src/components/home/search/RecentSearchContainer.tsx
@@ -1,5 +1,5 @@
 import { CloseIcon } from '@public/icon';
-import { SearchLogDtoDataTypes } from 'types/search/SearchTypes';
+import { SearchLogDtoDataTypes } from 'types/home/search/SearchTypes';
 import { useDeleteRecentSearch } from '@hooks/home/search/useDeleteRecentSearch';
 
 export interface RecentSearchContainerProps {
