+++ footlog/src/components/home/search/SearchingResults.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import RegionCard from '@components/home/search/RegionCard';
 import LocationCard from '@components/common/LocationCard/LocationCard';
-import { RegionCardDataTypes } from 'types/home/search/SearchTypes';
+import { CityRegionsDtoDataTypes } from 'types/home/search/SearchTypes';
 import { CourseResponseDtoDataTypes } from 'types/common/CommonTypes';
 
 interface SearchingResultsProps {
   filteredCourses: CourseResponseDtoDataTypes[];
-  filteredLocations: RegionCardDataTypes[];
+  filteredLocations: CityRegionsDtoDataTypes[];
   searchInput: string;
 }
 
@@ -18,7 +18,7 @@ export default function SearchingResults(props: SearchingResultsProps) {
       {filteredLocations.length > 0 && (
         <section className="mb-24pxr flex flex-col gap-24pxr">
           {filteredLocations.map((location) => (
-            <RegionCard key={location.id} location={location} searchInput={searchInput} />
+            <RegionCard key={location.sigunguCode} location={location} searchInput={searchInput} />
           ))}
         </section>
       )}
