+++ footlog/src/types/common/CommonTypes.ts
@@ -5,12 +5,6 @@ export interface NavBarIconsDataTypes {
   activeIcon: React.FC;
 }
 
-// 지역명, 최근 검색어, 인기 코스
-export interface RegionsDataTypes {
-  id: number;
-  name: string;
-}
-
 // 코스 검색 결과 LocationCard
 export interface LocationCardProps {
   course: CourseResponseDtoDataTypes;
