+++ footlog/src/types/home/search/SearchTypes.ts
@@ -1,11 +1,15 @@
-// 위치 검색 결과 RegionCard
-export interface RegionCardDataTypes {
-  id: number;
-  title: string;
-  subtitle?: string;
+// 위치 검색 결과
+export interface CityRegionsDtoDataTypes {
+  sigunguCode: number;
+  areaCode: number;
+  sigunguName: string;
 }
 
 export interface RegionCardProps {
-  location: RegionCardDataTypes;
+  location: CityRegionsDtoDataTypes;
   searchInput: string;
+}
+
+export interface SearchLogDtoDataTypes {
+  keyword: string;
 }
