--- footlog/src/types/search/SearchTypes.ts
@@ -1,3 +1,0 @@
-export interface SearchLogDtoDataTypes {
-  keyword: string;
-}
