+++ footlog/src/utils/filterData.ts
@@ -1,10 +1,10 @@
-import { RegionCardDataTypes } from 'types/home/search/SearchTypes';
+import { CityRegionsDtoDataTypes } from 'types/home/search/SearchTypes';
 import { CourseResponseDtoDataTypes } from 'types/common/CommonTypes';
 
 export const filterCourses = (courses: CourseResponseDtoDataTypes[], searchInput: string) => {
   return courses.filter((course) => course.name.toLowerCase().includes(searchInput.toLowerCase()));
 };
 
-export const filterLocations = (locations: RegionCardDataTypes[], searchInput: string) => {
-  return locations.filter((location) => location.title.toLowerCase().includes(searchInput.toLowerCase()));
+export const filterLocations = (locations: CityRegionsDtoDataTypes[], searchInput: string) => {
+  return locations.filter((location) => location.sigunguName.toLowerCase().includes(searchInput.toLowerCase()));
 };
